Ajouter champs image au model equipement

+++ Backend/src/main/java/support/backend/Repository/TechnicienRepository.java
@@ -3,5 +3,5 @@ package support.backend.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import support.backend.Model.Technicien;
 
-public interface TechnicienRepository extends JpaRepository<Technicien, Long> {
+public interface TechnicienRepository extends JpaRepository<Technicien, Integer> {
 }
